requirements.txt
Pillow==7.1.1
pyscreenshot==1.0
twilio==6.38.0